remove hashCode & equals from HAMT & RBT (#2187)
