+++ vavr/src/main/java/io/vavr/collection/RedBlackTree.java
@@ -312,22 +312,6 @@ interface RedBlackTree<T> extends Iterable<T> {
     }
 
     /**
-     * Compares color, value and sub-trees. The comparator is not compared because function equality is not computable.
-     *
-     * @return The hash code of this tree.
-     */
-    @Override
-    boolean equals(Object o);
-
-    /**
-     * Computes the hash code of this tree based on color, value and sub-trees. The comparator is not taken into account.
-     *
-     * @return The hash code of this tree.
-     */
-    @Override
-    int hashCode();
-
-    /**
      * Returns a Lisp like representation of this tree.
      *
      * @return This Tree as Lisp like String.
@@ -438,25 +422,6 @@ interface RedBlackTreeModule {
         @Override
         public T value() {
             return value;
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (o == this) {
-                return true;
-            } else if (o instanceof Node) {
-                final Node<?> that = (Node<?>) o;
-                return Collections.areEqual(this, that);
-            } else {
-                return false;
-            }
-        }
-
-        @Override
-        public int hashCode() {
-            // DEV-NOTE: Using `Objects.hash(this.value, this.left, this.right)` would leak the tree structure to the outside.
-            //           We just want to hash the values in the right order.
-            return Collections.hashOrdered(this);
         }
 
         @Override
@@ -912,17 +877,6 @@ interface RedBlackTreeModule {
         @Override
         public T value() {
             throw new NoSuchElementException("value on empty");
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            // note: it is not possible to compare the comparators because function equality is not computable
-            return (o == this) || (o instanceof Empty);
-        }
-
-        @Override
-        public int hashCode() {
-            return 1;
         }
 
         @Override
