+++ vavr/src/test/java/io/vavr/collection/HashArrayMappedTrieTest.java
@@ -143,100 +143,6 @@ public class HashArrayMappedTrieTest {
         assertThat(trie.get(null).get()).isEqualTo(2);  // key.hashCode = 0
     }
 
-    // -- equals
-
-    @SuppressWarnings("EqualsWithItself")
-    @Test
-    public void shouldEqualSameHAMTInstance() {
-        final HashArrayMappedTrie<Integer, Integer> trie = empty();
-        assertThat(trie.equals(trie)).isTrue();
-    }
-
-    @Test
-    public void shouldEmptyNotEqualsDifferentType() {
-        assertThat(empty().equals("")).isFalse();
-    }
-
-    @Test
-    public void shouldNonEmptyNotEqualsDifferentType() {
-        assertThat(of(1).equals("")).isFalse();
-    }
-
-    @Test
-    public void shouldRecognizeEqualityOfNils() {
-        assertThat(empty().equals(empty())).isTrue();
-    }
-
-    @Test
-    public void shouldRecognizeEqualityOfNonNils() {
-        assertThat(of(1, 2, 3).equals(of(1, 2, 3))).isTrue();
-    }
-
-    @Test
-    public void shouldRecognizeNonEqualityOfHAMTOfSameSize() {
-        assertThat(of(1, 2, 3).equals(of(1, 2, 4))).isFalse();
-    }
-
-    @Test
-    public void shouldRecognizeNonEqualityOfHAMTOfDifferentSize() {
-        assertThat(of(1, 2, 3).equals(of(1, 2))).isFalse();
-        assertThat(of(1, 2).equals(of(1, 2, 3))).isFalse();
-    }
-
-    @Test
-    public void shouldEqualsIgnoreOrder() {
-        HashArrayMappedTrie<String, Integer> map = HashArrayMappedTrie.<String, Integer> empty().put("Aa", 1).put("BB", 2);
-        HashArrayMappedTrie<String, Integer> map2 = HashArrayMappedTrie.<String, Integer> empty().put("BB", 2).put("Aa", 1);
-        assertThat(map.hashCode()).isEqualTo(map2.hashCode());
-        assertThat(map).isEqualTo(map2);
-    }
-
-    @Test
-    public void shouldCalculateHashCodeOfLongLeafList() {
-        HashArrayMappedTrie<WeakInteger, Integer> h1 = HashArrayMappedTrie.empty();
-        for (int i = 0; i < 100000; i++) {
-            h1 = h1.put(new WeakInteger(i), i);
-        }
-        assertThat(h1.hashCode()).isNotZero();
-    }
-
-    // -- hashCode
-
-    @Test
-    public void shouldCheckHashCodeInLeafList() {
-        HashArrayMappedTrie<Integer, Integer> trie = empty();
-        trie = trie.put(0, 1).put(null, 2);       // LeafList.hash == 0
-        final Option<Integer> none = trie.get(1 << 6);  // (key.hash & BUCKET_BITS) == 0
-        assertThat(none).isEqualTo(Option.none());
-    }
-
-    @Test
-    public void shouldCalculateHashCodeOfNil() {
-        assertThat(empty().hashCode()).isEqualTo(1);
-    }
-
-    @Test
-    public void shouldCalculateHashCodeOfCollision() {
-        assertThat(empty().put(null, 1).put(0, 2).hashCode()).isEqualTo(empty().put(null, 1).put(0, 2).hashCode());
-    }
-
-    @Test
-    public void shouldCalculateDifferentHashCodesForDifferentHAMT() {
-        assertThat(of(1, 2).hashCode()).isNotEqualTo(of(2, 3).hashCode());
-    }
-
-    @Test
-    public void shouldCalculateBigHashCode() {
-        HashArrayMappedTrie<Integer, Integer> h1 = empty();
-        HashArrayMappedTrie<Integer, Integer> h2 = empty();
-        final int count = 1234;
-        for (int i = 0; i <= count; i++) {
-            h1 = h1.put(i, i);
-            h2 = h2.put(count - i, count - i);
-        }
-        assertThat(h1.hashCode() == h2.hashCode()).isTrue();
-    }
-
     // - toString
 
     @Test
