+++ vavr/src/test/java/io/vavr/collection/HashMapTest.java
@@ -20,6 +20,8 @@
 package io.vavr.collection;
 
 import io.vavr.Tuple2;
+import io.vavr.control.Option;
+import org.assertj.core.api.Assertions;
 import org.junit.Test;
 
 import java.math.BigDecimal;
@@ -136,6 +138,44 @@ public class HashMapTest extends AbstractMapTest {
         source.put(1, 2);
         source.put(3, 4);
         assertThat(HashMap.ofAll(source)).isEqualTo(emptyIntInt().put(1, 2).put(3, 4));
+    }
+
+    // -- specific
+
+    @Test
+    public void shouldCalculateHashCodeOfCollision() {
+        Assertions.assertThat(HashMap.empty().put(null, 1).put(0, 2).hashCode())
+                .isEqualTo(HashMap.empty().put(0, 2).put(null, 1).hashCode());
+        Assertions.assertThat(HashMap.empty().put(null, 1).put(0, 2).hashCode())
+                .isEqualTo(HashMap.empty().put(null, 1).put(0, 2).hashCode());
+    }
+
+    @Test
+    public void shouldCheckHashCodeInLeafList() {
+        HashMap<Integer, Integer> trie = HashMap.empty();
+        trie = trie.put(0, 1).put(null, 2);       // LeafList.hash == 0
+        final Option<Integer> none = trie.get(1 << 6);  // (key.hash & BUCKET_BITS) == 0
+        Assertions.assertThat(none).isEqualTo(Option.none());
+    }
+
+    @Test
+    public void shouldCalculateBigHashCode() {
+        HashMap<Integer, Integer> h1 = HashMap.empty();
+        HashMap<Integer, Integer> h2 = HashMap.empty();
+        final int count = 1234;
+        for (int i = 0; i <= count; i++) {
+            h1 = h1.put(i, i);
+            h2 = h2.put(count - i, count - i);
+        }
+        Assertions.assertThat(h1.hashCode() == h2.hashCode()).isTrue();
+    }
+
+    @Test
+    public void shouldEqualsIgnoreOrder() {
+        HashMap<String, Integer> map = HashMap.<String, Integer> empty().put("Aa", 1).put("BB", 2);
+        HashMap<String, Integer> map2 = HashMap.<String, Integer> empty().put("BB", 2).put("Aa", 1);
+        Assertions.assertThat(map.hashCode()).isEqualTo(map2.hashCode());
+        Assertions.assertThat(map).isEqualTo(map2);
     }
 
     // -- spliterator
